Our local node height is actually a uint64_t if we believe with the daemon RPC messages tells us

+++ src/WalletService/PaymentServiceJsonRpcMessages.h
@@ -95,7 +95,7 @@ struct GetStatus {
   struct Response {
     uint32_t blockCount;
     uint32_t knownBlockCount;
-    uint32_t localDaemonBlockCount;
+    uint64_t localDaemonBlockCount;
     std::string lastBlockHash;
     uint32_t peerCount;
